+++ tests/Adapter/LoggerTest.php
@@ -8,6 +8,7 @@ use Innmind\Filesystem\{
     Adapter,
     File,
     Name,
+    Directory\Directory,
 };
 use Innmind\Immutable\{
     Set,
@@ -122,12 +123,18 @@ class LoggerTest extends TestCase
             $inner = $this->createMock(Adapter::class),
             $this->createMock(LoggerInterface::class),
         );
-        $all = Set::of(File::class);
+        $all = Set::of($file = File\File::named(
+            'watev',
+            $this->createMock(File\Content::class),
+        ));
         $inner
             ->expects($this->once())
-            ->method('all')
-            ->willReturn($all);
+            ->method('root')
+            ->willReturn(Directory::of(
+                Name::of('root'),
+                $all,
+            ));
 
-        $this->assertSame($all, $adapter->all());
+        $this->assertSame([$file], $adapter->all()->toList());
     }
 }